+++ core/utils/lume_config.ts
@@ -13,6 +13,7 @@ export const pluginNames = [
   "extract_date",
   "extract_order",
   "git_date",
+  "git_info",
   "jsx",
   "json_ld",
   "reading_info",